Re-org delete db views
