--- packages/admin-ui/src/pages/system/components/Advanced/CacheManager/index.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import Card from "components/Card";
-import ClearCache from "../CacheManager/ClearCacheManager";
-
-export function ClearCacheManager() {
-  return (
-    <Card spacing>
-      <ClearCache />
-    </Card>
-  );
-}
