--- packages/admin-ui/src/pages/system/components/Advanced/MaindbManager/index.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import Card from "components/Card";
-import ClearMainDb from "./ClearMainDb";
-
-export function MaindbManager() {
-  return (
-    <Card spacing>
-      <ClearMainDb />
-    </Card>
-  );
-}
